% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lineages.R
\name{get_lineages}
\alias{get_lineages}
\title{Get lineages}
\usage{
get_lineages(tree, depth)
}
\arguments{
\item{tree}{phylo class object}

\item{depth}{depth, in same units as branch length, at which lineages will be obtained}
}
\value{
data frame of with columns 'species' containing all species in the tree and "ID" indicating the lineage ID a species is in.
}
\description{
Get lineages
}
\examples{
library(phylosympatry)
tree <- rtree(100)
get_lineages(tree,.3)
get_lineages(tree,1)
}
