% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_collapse.R
\name{lineage_collapse}
\alias{lineage_collapse}
\title{Collapse lineages in presence-matrix}
\usage{
lineage_collapse(lineages, presence_matrix, collapse.fcn = max)
}
\arguments{
\item{lineages}{lineages object, see \code{\link{get_lineages}}. Must have columns \code{species} and \code{ID}, where all species are present in the rownames of the presence matrix}

\item{presence_matrix}{Matrix whose rownames contain all the species in \code{lineages}}

\item{collapse.fcn}{Aggregatin function for collapsing lineages. Takes as input vectors of data (may have length 1) from presence_matrix and outputs numerical value. Default is \code{\link{max}}}
}
\value{
matrix of numeric values output from \code{collapse.fcn} with number of rows equal to \code{length(unique(lineages$ID))}, and columns equal to \code{ncol(presence_matrix)}
}
\description{
Collapse lineages in presence-matrix
}
\examples{
library(phylosympatry)
set.seed(1)
m=20 #number of species
n=10  #number of sites
tree <- ape::rtree(m)
lineages <- get_lineages(tree,1)
presence_matrix <- matrix(rbinom(m*n,1,.1),nrow=m,ncol=n)
rownames(presence_matrix) <- tree$tip.label
colnames(presence_matrix) <- letters[1:n]

Mat <- lineage_collapse(lineages,presence_matrix)
}
