% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineage_regression.R
\name{lineage_regression}
\alias{lineage_regression}
\title{Regression of covariate, x, on lineage data in presence matrix, aggregated via collapse.fcn}
\usage{
lineage_regression(lineages, presence_matrix, x = "richness",
  collapse.fcn = max, ncores = NULL, ...)
}
\arguments{
\item{lineages}{lineages object output from \code{\link{get_lineages}}, with columns \code{species} and \code{ID}}

\item{presence_matrix}{Matrix whose rownames contain all the species in \code{lineages}}

\item{x}{covariate of length equal to \code{ncol(presence_matrix)}. Default is the richness, or \code{colSums(presence_matrix)}}

\item{collapse.fcn}{Aggregatin function for collapsing lineages. Takes as input vectors of data (may have length 1) from presence_matrix and outputs numerical value. Default is \code{\link{max}}}

\item{ncores}{optional input number of cores for parallelized computation of many glm objects. May be memory-intensive with large \code{presence_matrix}}
}
\value{
named list, with each element being a generalized linear model \code{glm(y~x,...)} where \code{y} is the collapsed data for a given lineage
}
\description{
Regression of covariate, x, on lineage data in presence matrix, aggregated via collapse.fcn
}
\examples{
library(phylosympatry)
set.seed(1)
m=20 #number of species
n=10  #number of sites
tree <- ape::rtree(m)
lineages <- get_lineages(tree,1)
presence_matrix <- matrix(rbinom(m*n,1,.1),nrow=m,ncol=n)
rownames(presence_matrix) <- tree$tip.label
colnames(presence_matrix) <- letters[1:n]
GLMs <- lineage_regression(lineages,presence_matrix)
}
