% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Richness_stats.R
\name{Richness_stats}
\alias{Richness_stats}
\title{Richness statistics for multiple communities}
\usage{
Richness_stats(assigned_lineages, presence_matrix)
}
\arguments{
\item{assigned_lineages}{output from \code{\link{lineage_assign}}}

\item{presence_matrix}{presence_absence matrix of distributional data, species names in the rownames}
}
\description{
Richness statistics for multiple communities
}
\examples{
library(phylosympatry)
set.seed(1)
m=100
tree <- ape::rtree(m)
lineages <- get_lineages(tree,1)
assigned_lineages <- lineage_assign(lineages,runif(nrow(lineages)),.7,.2,.1,.3)

n=5
presence_matrix <- matrix(rbinom(m*n,1,.1),nrow=m)
rownames(presence_matrix) <- tree$tip.label

Richness_stats(assigned_lineages,presence_matrix)
}
