% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ricklefs_analysis.R
\name{ricklefs_analysis}
\alias{ricklefs_analysis}
\title{Compute whether lineages that characteristically occur at species-rich sites have different average richness that lineages that characteristically occur at species-poor sites}
\usage{
ricklefs_analysis(assigned_lineages)
}
\arguments{
\item{assigned_lineages}{output from \code{\link{lineage_assign}}}
}
\description{
Compute whether lineages that characteristically occur at species-rich sites have different average richness that lineages that characteristically occur at species-poor sites
}
\examples{
library(phylosympatry)
tree <- ape::rtree(100)
lineages <- get_lineages(tree,1)
species_prop <- runif(nrow(lineages))
assigned_lineages <- lineage_assign(lineages,species_prop,.7,.1,.3,.2)
ricklefs_analysis(assigned_lineages)
}
